코드 상속
