+++ Assets/Prefab/Enemy/GrassGolem.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 6520cac900ac7534ba261ab28c6cea99
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
