+++ Assets/Prefab/Enemy/WaterGolem.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a63a6edbf7516244f8ce63b5a3bf4f99
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
