+++ Assets/ScriptableObjects/Enemy.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7e2d385083cff2842b43259209f98e63
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
