+++ Assets/ScriptableObjects/Enemy/GrassGolem.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 03fa06eaac043fb459e1e667341fe3d3
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
