+++ Assets/ScriptableObjects/Enemy/WaterGolem.asset
@@ -1,0 +1,23 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 80a29fc66d98ff94696da7dc348f00a6, type: 3}
+  m_Name: WaterGolem
+  m_EditorClassIdentifier: 
+  Score: 50
+  MinX: -15
+  MaxX: 15
+  MoveSpeed: 5
+  ShootCooldown: 10
+  RandomShootCooldown: 5
+  projectile: {fileID: -8760447939717194727, guid: 7c8838ee1d309864c94668d388447f1d, type: 3}
+  DamageParticle: {fileID: 1590039965784815200, guid: 567a9eba6cc906947beccf494bfe08c3, type: 3}
+  YOffset: 0
