+++ Assets/ScriptableObjects/Enemy/WaterGolem.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f5cf4b20bdfa9054d9c22c27187b8b71
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
